% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_get_url.R
\name{xlsx_get_url}
\alias{xlsx_get_url}
\title{Extract URLs from Excel xlsx hyperlinks}
\usage{
xlsx_get_url(xlsx_file, sheet = 1)
}
\arguments{
\item{xlsx_file}{a string indicating the xlsx file's path.}

\item{sheet}{an inetger or a string indicating the xlsx sheet.}
}
\value{
a tibble.
}
\description{
Extract URLs from Excel xlsx hyperlinks
}
